Avoid trying to parser log lines when parser is not initialized
Currently, the collector log processing code assumes that the log
parser has always been initialized by the time we receive log messages
to process.  This is generally true, but it can be false if the
collector failed to connect to the database (to fetch log_line_prefix
to initialize the parser). That can lead to confusing errors (that
are not relevant to the root cause of the connection problem).

Skip attempting to process log lines when the log parser is not yet
initialized.