Pass buffercache statistics in statistics state structs
These were located in the structs that are typically used for metadata
about the table, with statistics kept separate, even if the statistics
in question are not diffed.

This does not make much of a difference in practice in the current way
this data is utilized (its merely a matter of following the existing
pattern), but will enable a subsequent refactoring of the partition
statistics merging logic, since it can rely on the statistics state
information alone (which is indexed by OID).