Introduce collection helper to pass secondary variables in input logic
The functions collecting data in input/postgres often need auxilliary
information like the Postgres version, the global collection options,
or the server configuration. They also sometimes need access to the
self test utilities, or the logger, which need to be passed along as well.

This has historically caused a lot of churn in the argument lists
(which grow very long), as well as inconsistencies in the ordering.

To unify all of the above, introduce the CollectionHelper, which is
intended to be passed as the second argument (after context) in all
input/postgres collection functions.