Improve helper function checks for consistency and lower overhead
This moves the checking for the presence of helper functions to the
collection helper, and relies on an upfront querying of pg_proc, instead
of querying for each helper individually. Note that we have to support
this both for the main database (for all server-wide data points), as
well as for each database to get column/extended statistics information.

In passing we also improves the support for checking for different
(potentially overloaded) versions of the same functions, by requiring
passing the function argument types to the lookup. That is specifically
necessary for the multiple historic variants of get_stat_statements that
may be present on the same database server, and lets us remove special
casing that code.