GH release action: Fix "packages" name conflict
During the release process we group together the created packages as
an artifact, to then upload them individually to the GitHub release entry.

Previously this was able to use a single "packages" zip file for both
x86_64 and arm64 packages, but we now need to split this since we split
the actions.