Update messages when the monitoring user doesn't have a proper setup (#699)
Previously, when the monitoring user is neither superuser nor has
pg_monitor role, we were showing error messages saying that please use a
superuser or create a helper function. While this is true for Aiven,
especially for "creating a helper function" guidance, it's not true for
other cases and when the user hits this with other providers, the error
message is confusing and make it not clear which actions to take.
Update the messaging to make it clear the next action for these cases.