Better support for running EXPLAIN with pganalyze.explain_analyze helper
With more tests, we noticed that passing parameter values and types
separately to this function is not working well for several cases.
For values, as `pg_catelog.string_agg` will never emit NULLs and only
aggreagates non-null values, it fails when we actually want to pass NULL
as a value. Use array_to_string instead as a workaround.
For types, when something like int4[] is passed, as it was quoting as
"int4[]" previously, this also didn't work. We check the input type and
if it matches to the regexp (type pattern, especially ones the pganalyze
app would pass), it skips quote_ident.