Amazon RDS: Also check monitoring interval for Enhanced Monitoring
We previously only checked for the Enhanced Monitoring Resource ARN to
be set, but that field will remain set when Enhanced Monitoring was once
turned on, but then turned off. Instead also check the monitoring interval
which will be 0 in such situations.

Fixes missing system metrics on RDS when Enhanced Monitoring is turned off.