Updated cloudsql module and defaulted to using private IP. (#704)
* Updated cloudsql module and defaulted to using private IP.

The CloudSQL driver discovers the IP address for an instance based on the project, instance, and region settings. If public IP access is not available, we must specify to use the private IP, which is presumably more often the default.

Added an option to use the public IP if needed.

* Address review feedback

---------

Co-authored-by: Ryan Booz <ryan@pganalyze.com>