Rework pg_stat_statements fetching to separate text vs stats
Since we now load one or the other (instead of both) we can invent
special purpose methods for each, avoiding a bunch of showtext checks
and making the logic easier to follow.

In passing, remove a nonsensical check for sql.ErrNoRows on the result
of db.PrepareContext. This only occurs when using QueryRow in Go, not
when querying for multiple rows.