Avoid skipping query stats in case of slow query text/schema collection (#730)
* Avoid skipping query stats in case of slow query text/schema collection

Previously the high-frequency stats collection (which occurs every minute,
except at the 10 minute mark) shared a mutex with the full snapshot
collection, in order to record query stats. This effectively meant that
query stats would be missed if the full snapshot took longer than 1
minute, which was observed in practice in the case of very slow query
text file loading (up to 2 minutes by default, higher with changed
settings), or when collecting schema data from many databases (up to 8
minutes total).

Instead, give the high frequency collection its own persisted state
struct and mutex, and move data from that struct during the full
snapshot. Additionally, to improve correctness of statistics, start
each full snapshot by calling the exact same logic that otherwise runs
once a minute. Perform the query text collection afterwards, without
recording statistics at that time.

This effectively means that we are doing an additional pg_stat_statements
access (without query text), but since that only looks at shared memory
and holds the pgss lock very briefly, it should not be a problem, and
avoids incorrect statistics when query text collection is slow (since
pg_stat_statements looks at the statistics after accessing the query text
file).

In passing, improve the correctness of the next query stats collection
after a reset (it had the wrong reference point before), and add a
missing mutex to the state file write (which now also writes out the
separate high frequency state).

An existing short-coming is left for a future improvement (but with an
added note), which is that state file updates only get written every 10
minutes, missing high-frequency updates inbetween in case of crashes.

* Rework pg_stat_statements fetching to separate text vs stats

Since we now load one or the other (instead of both) we can invent
special purpose methods for each, avoiding a bunch of showtext checks
and making the logic easier to follow.

In passing, remove a nonsensical check for sql.ErrNoRows on the result
of db.PrepareContext. This only occurs when using QueryRow in Go, not
when querying for multiple rows.

* CollectFull: Collect some metrics earlier that are fast to collect

This improves the correctness of replication/server/backend count
statistics, as they are collected closer to the reported collected_at
timestamp.

To help future modifications in the method, add a comment to the function
explaining how to choose when to collect data.