Refactor legacy HTTP uploading code to be better encapsulated
This moves the legacy HTTP upload+submit logic to be hidden behind a
separate Go routine (which decides whether to use a WebSocket, or fall
back to HTTP). Output methods now always send snapshots to dedicated
channels per snapshot type.

Legacy HTTP handling is moved to a separate file to make it clear when
it is utilized.

In passing, introduce a helper to get the snapshot API headers, and
pass protobuf snapshot structs by reference in more places, to avoid
a Go warning related to copying a contained mutex.