Refactor WebSocket handling into a separate utility, start only if needed
This introduces a new helper, the "reconnecting socket", which maintains
a WebSocket connection and can reconnect as needed. This abstracts
existing WebSocket behaviour better, and allows starting the WebSocket
only when first used, instead of when initializing it.

With this change, we only start the WebSocket connection when we start
a snapshot and need a grant. Additionally, this introduces a wait of up
to 1 second to allow the server to send the initial configuration message
after connecting. If it takes longer the initial grant will utilize an
HTTP-based grant (this was already the case before in certain cases).

Additionally this changes the reconnection interval: Before we always
reconnected every 3 minutes if needed. Now we reconnect every 1 min,
unless the connection failed with a 4XX error, in which case we delay
for 8 minutes. This intends to skip over authentication issues with
expired API keys, whilst allowing for misconfigured Enterprise Server
installs (so the next full snapshot recovers, if needed).