Helper: Don't require "locate" to determine the pg_controldata path
Previously we relied on the locate utility to find pg_controldata on
the filesystem, but that can be brittle and doesn't work when locate
is not installed, like it was observed on Debian/Ubuntu systems, that
also don't have Postgres binaries in the PATH with the PGDG packages.

Instead, use the directory the postmaster executable is located in
as the assumed location of the pg_controldata binary.

This fixes cluster ID detection (via the Postgres system identifier)
for a typical PGDG-based Debian/Ubuntu install.