Add new configuration setting to require WebSockets, default off for now (#745)
This introduces a new setting "api_require_websocket" /
"API_REQUIRE_WEBSOCKET" that requires the Snapshot API connection to
go over the newer WebSocket API instead of the legacy HTTPS API.

The WebSocket API is geared towards long-lived connections (lower overhead)
and allows the server side to push down requests, like for running
EXPLAIN on-demand as part of the opt-in collector workflow for Workbooks.

We are currently planning to change the default value of this to "on"
in a future collector release. Currently this defaults to "off" which is
the same behaviour as before (try WebSocket, but allow HTTPS fallback).