OTel Log Server: Support receiving JSON log format in addition to Protobuf (#747)
* OTel Log Server: Support receiving JSON log format in addition to Protobuf

This supports forwarding logs using Vector's OpenTelemetry Logs sink when
configured to map the record information correctly.

With thanks to Rauan Mayemir for contributing the initial code to fix
this.

* Skip log_line_prefix validation when using OpenTelemetry Logs server

We currently assume that CloudNativePG is always used when utilizing
the OTel Logs server, and CNPG always emits jsonlog, which reliably has
the username/database as structured information.

* Improve log messages to prefix them with "OTel log server" consistently