Rework OTel logs handler to be spec compliant, refactor for clarity
The OTLP protocol spec requires that we return a response object
(encoded based on request content type) that indicates how many of the
submitted log records were rejected. Further it requires us to return a
4XX error if we could not process the request. Any error thats not 429
(rate limited) must not be retried by the client, per the spec.

In passing, refactor the OTel logs handler to reduce duplication when
dealing with very similar yet different cases (jsonlog as the main log
record, vs wrapped in Kubernetes context).